requirements.txt
vosk>=0.3.45
argostranslate>=1.9.0
piper-tts>=1.2.0
pyalsaaudio>=0.10.0
flask>=3.0.0
pyyaml>=6.0

# dev dependencies
pytest>=8.0.0
pytest-cov>=5.0.0
